matched sub_8048A68 thanks to @YohannDR
